Add files via upload
A registration and login system would be the basic CTA for a website that you can try building using Django. The registration page would include fields such as the mail-id or phone number, the username(optional), new password and a remember me checkbox. The user can sign up with their mail-id or phone number and set a password according to the guidelines. The profile is activated via an activation code which is mailed to the user.

Later, the login page displays fields to enter the username/mail-id/phone number and password. The user can then click on the login button to log into the website.
Additionally, you can also implement a password reset mechanism.